[FIX] Deleted test specific string file, fixed some linter warnings
