+++ src/test_net_propagation.py
@@ -2,7 +2,6 @@
 
 import net_propagation
 import strings
-import test_files.strings
 
 """
  - Importing net_propagation for testing.
@@ -27,8 +26,8 @@ def test_additional_actions():
     arguments = [strings.ARGUMENT_IP_ADDRESS_FILENAME,
                  strings.ARGUMENT_SPECIFIC_PROPAGATION_FILE]
     ip = strings.BLANK_IP
-    username = test_files.strings.RANDOM_STRING
-    transfer_file_filename = test_files.strings.RANDOM_STRING
+    username = strings.RANDOM_STRING
+    transfer_file_filename = strings.RANDOM_STRING
     ports = [strings.SSH_PORT, strings.TELNET_PORT,
              strings.WEB_PORT_EIGHTY,
              strings.WEB_PORT_EIGHTY_EIGHTY,
@@ -46,14 +45,14 @@ def test_append_lines_from_file_to_list():
     for readability in test results i.e. we'll be able to correlate a specific
     line with an error.
     """
-    with open(str(test_files.strings.TEST_FILENAME)) as file:
+    with open(str(strings.FILE)) as file:
         lines_list = net_propagation.append_lines_from_file_to_list(file)
-    assert lines_list[0] == test_files.strings.TEST_LINES[0]
-    assert lines_list[1] == test_files.strings.TEST_LINES[1]
-    assert lines_list[2] == test_files.strings.TEST_LINES[2]
-    assert lines_list[3] == test_files.strings.TEST_LINES[3]
-    assert lines_list[4] == test_files.strings.TEST_LINES[4]
-    assert lines_list[5] == test_files.strings.TEST_LINES[5]
+    assert lines_list[0] == strings.LINES[0]
+    assert lines_list[1] == strings.LINES[1]
+    assert lines_list[2] == strings.LINES[2]
+    assert lines_list[3] == strings.LINES[3]
+    assert lines_list[4] == strings.LINES[4]
+    assert lines_list[5] == strings.LINES[5]
 
 
 def test_assigning_values():
@@ -63,7 +62,6 @@ def test_assigning_values():
     before it does that the bad path is checked by passing in a single argument
     with no value to get a runtime error.
     """
-
 
 
 def test_exit_and_show_instructions(capfd):
